+++ packages/observer/src/intersection-observer.d.ts
@@ -1,0 +1,10 @@
+/**
+ * intersection-observer polyfill 模块声明
+ *
+ * 这个模块是一个 polyfill，不需要导出任何内容
+ * 它会在全局作用域中定义 IntersectionObserver
+ */
+declare module "intersection-observer" {
+	const _: undefined;
+	export = _;
+}